+++ NEWS.md
@@ -1,3 +1,16 @@
+# GiottoClass 0.5.0
+
+## changes
+- `calculateOverlap()` and `overlapToMatrix()` param harmonization
+
+## new
+- `aggregateFeatures()` wrapper for running `calculateOverlap()` and `overlapToMatrix()`
+- `overlapPointDT()` and `overlapIntensityDT()` classes to store overlaps relationships efficiently and help with aggregation pipeline
+
+## bug fixes
+- `overlaps()` will now properly find image overlaps
+
+
 # GiottoClass 0.4.8 (2025/06/16)
 
 ## new